requirements.txt
chainlit
langchain-ollama
langchain-chroma
langchain-core
ollama
docx2txt
pypdf
openai-whisper
python-dotenv
PyYAML
